move splitstore into blockstore package.
